requirements.txt
six==1.11.0
future==0.16.0
coverage==4.4.2
codecov==2.0.10
django==1.11.8
django-dotenv==1.4.2
django-debug-toolbar==1.9.1
django-widget-tweaks==1.4.1
django-storages==1.6.5
djangorestframework==3.7.3
django-filter==1.1.0
django-maintenance-mode==0.7.0
django-pipeline==1.6.13
celery[redis]==4.1.0
django-redis==4.8.0
statsd==3.2.1
boto==2.48.0
pyOpenSSL==17.5.0
pycrypto==2.6.1
Pillow==2.5.3
icalendar==4.0.0
python-dateutil==2.6.1
mock==2.0.0
mysqlclient==1.3.12
coreapi==2.3.3
Pygments==2.2.0
Markdown==2.6.10
Sphinx==1.6.5